adding playRound to game() loop
Was able to succesfully get the playRound function
within the game() function to work. Had to consult some
of the Odin Project users on Discord for advice. I had created
a shoot() function to guess a random value of RPS from an array
and the I passed it through another function getComputerChoice().
This would always reutnr undefined, and I realized 2 options. 1 is
that you could nix getComputerChoice() and just use shoot, or you
need to include the return value for Shoot() within the gCC function
to carry along the string value or else gCC is undefined when it
makes its way into the playRound() function.