cleaning up RPS game
I went through and cleaned up the code, as well
as went through and put comments in correct areas,
while deleting out the vast majority of the pseudocode.
Put the pseudocode into HTML for now. When it comes time to
revisit, will likely recycle pseudocode, or keep as a seperate
HTML page for reference on how to improve.